+++ app/services/push_notifications.py
@@ -1,0 +1,96 @@
+import logging
+from datetime import datetime, timezone
+
+from app.helpers.channels import get_channel_online_users, get_channel_users, is_user_in_channel
+from app.helpers.events import EventType
+from app.helpers.expo import send_expo_push_notifications
+from app.helpers.list_utils import batch_list
+from app.helpers.message_utils import get_message_mentions, get_raw_blocks
+from app.models.channel import ChannelReadState
+from app.models.message import Message
+from app.models.user import User
+from app.schemas.channels import ChannelReadStateCreateSchema
+from app.services.crud import create_item, find_and_update_item, get_item_by_id
+
+logger = logging.getLogger(__name__)
+
+
+async def dispatch_push_notification_event(event: EventType, data: dict):
+    if event != EventType.MESSAGE_CREATE:
+        return
+
+    message_dict = data.get("message", {})
+    app_dict = data.get("app", {})
+    message_id = message_dict.get("id")
+    message = await get_item_by_id(id_=message_id, result_obj=Message)
+    channel = await message.channel.fetch()
+    if not channel:
+        raise Exception("expected message to have a channel")
+
+    if message.author:
+        author: User = await message.author.fetch()
+        author_name = author.display_name or author.wallet_address
+    elif app_dict:
+        author_name = app_dict.get("name")
+    else:
+        author_name = "New message"
+
+    push_title = f"{author_name} (#{channel.name})"
+    push_body = (await get_raw_blocks(message.blocks))[:100]
+    push_metadata = {**data, "event": event.name, "url": f"channels/{str(channel.pk)}"}
+    push_data = {"title": push_title, "body": push_body, "data": push_metadata, "sound": "default"}
+
+    users_to_notify = []
+    mentions = await get_message_mentions(message)
+    for mention_type, mention_ref in mentions:
+        logger.debug(f"should send notification of type '{mention_type}' to @{mention_ref}")
+
+        if mention_type == "user":
+            user_id = mention_ref
+            user = await get_item_by_id(id_=user_id, result_obj=User)
+            users_to_notify.append(user)
+
+        if mention_type == "broadcast":
+            if mention_ref == "here":
+                online_users = await get_channel_online_users(channel=channel)
+                users_to_notify.extend(online_users)
+            elif mention_ref == "channel" or mention_ref == "everyone":
+                channel_users = await get_channel_users(channel=channel)
+                users_to_notify.extend(channel_users)
+            else:
+                logger.warning(f"unknown mention reference: {mention_ref}")
+                continue
+
+    if len(users_to_notify) == 0:
+        logger.debug("no users to push notifications")
+        return
+
+    push_messages = []
+
+    for user in users_to_notify:
+        user_in_channel = await is_user_in_channel(user=user, channel=channel)
+        if not user_in_channel:
+            continue
+
+        read_state = await find_and_update_item(
+            filters={"user": user.pk, "channel": channel.pk},
+            data={"$inc": {"mention_count": 1}},
+            result_obj=ChannelReadState,
+        )
+
+        if read_state:
+            mention_count = read_state.mention_count
+        else:
+            mention_count = 1
+            read_state_model = ChannelReadStateCreateSchema(
+                channel=str(channel.id),
+                last_read_at=datetime.fromtimestamp(0, tz=timezone.utc),
+                mention_count=mention_count,
+            )
+            await create_item(read_state_model, result_obj=ChannelReadState, current_user=user)
+
+        if user.push_tokens and len(user.push_tokens) > 0:
+            push_messages.append({**push_data, "to": user.push_tokens, "badge": mention_count})
+
+    async for batched_messages in batch_list(push_messages):
+        await send_expo_push_notifications(push_messages=batched_messages)
